SuperRay-Backend/requirements.txt
absl-py==2.0.0
addict==2.4.0
asgiref==3.7.2

cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

filelock==3.13.1
filterpy==1.4.5
fonttools==4.48.1
fsspec==2023.12.1
future==0.18.3
google-auth==2.25.1
google-auth-oauthlib==1.1.0
grpcio==1.60.0
gunicorn==20.1.0
idna==3.6
imageio==2.33.0
importlib-metadata==7.0.0
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
lmdb==1.4.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.2.1
Pillow==10.1.0
platformdirs==4.1.0
protobuf==4.25.1
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12




tifffile==2023.9.26
tomli==2.0.1

tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.1.27
urllib3==2.1.0
Werkzeug==3.0.1
whitenoise==6.6.0
yapf==0.40.2
zipp==3.17.0
